requirements.txt
appdirs==1.4.4
asgiref==3.3.0
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.1
Pillow==9.0.0
pyparsing==2.4.7
pytz==2020.4
regex==2020.10.28
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.5.4
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11

djangorestframework-jwt~=1.11.0